@@ -3,82 +3,102 @@ package upstream
 import (
 	"context"
 	"fmt"
-	"net"
+	"net/netip"
 	"time"
 
 	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/miekg/dns"
+	"golang.org/x/exp/slices"
 )
 
-// exchangeResult is a structure that represents result of exchangeAsync
-type exchangeResult struct {
-	reply    *dns.Msg // Result of DNS request execution
-	upstream Upstream // Upstream that successfully resolved request
-	err      error    // Error
-}
+const (
+	// ErrNoUpstreams is returned from the methods that expect at least a single
+	// upstream to work with when no upstreams specified.
+	ErrNoUpstreams errors.Error = "no upstream specified"
 
-// ErrNoUpstreams is returned from the methods that expect at least a single
-// upstream to work with when no upstreams specified.
-const ErrNoUpstreams errors.Error = "no upstream specified"
+	// ErrNoReply is returned from [ExchangeAll] when no upstreams replied.
+	ErrNoReply errors.Error = "no reply"
+)
 
-// ExchangeParallel function is called to parallel exchange dns request by many upstreams
-// First answer without error will be returned
-// We will return nil and error if count of errors equals count of upstreams
-func ExchangeParallel(u []Upstream, req *dns.Msg) (*dns.Msg, Upstream, error) {
-	size := len(u)
-
-	if size == 0 {
+// ExchangeParallel returns the dirst successful response from one of u.  It
+// returns an error if all upstreams failed to exchange the request.
+func ExchangeParallel(ups []Upstream, req *dns.Msg) (reply *dns.Msg, resolved Upstream, err error) {
+	upsNum := len(ups)
+	switch upsNum {
+	case 0:
 		return nil, nil, ErrNoUpstreams
+	case 1:
+		reply, err = exchangeAndLog(ups[0], req)
+
+		return reply, ups[0], err
+	default:
+		// Go on.
 	}
 
-	if size == 1 {
-		reply, err := exchange(u[0], req)
-		return reply, u[0], err
-	}
-
-	// Size of channel must accommodate results of exchangeAsync from all upstreams
-	// Otherwise sending in channel will be locked
-	ch := make(chan *exchangeResult, size)
-
-	for _, f := range u {
-		go exchangeAsync(f, req, ch)
+	resCh := make(chan any, upsNum)
+	for _, f := range ups {
+		go exchangeAsync(f, req, resCh)
 	}
 
 	errs := []error{}
-	for n := 0; n < len(u); n++ {
-		rep := <-ch
-		if rep.err != nil {
-			errs = append(errs, rep.err)
-		} else if rep.reply != nil {
-			return rep.reply, rep.upstream, nil
+	for range ups {
+		var r *ExchangeAllResult
+		r, err = receiveAsyncResult(resCh)
+		if err != nil {
+			if !errors.Is(err, ErrNoReply) {
+				errs = append(errs, err)
+			}
+		} else {
+			return r.Resp, r.Upstream, nil
 		}
 	}
 
+	// TODO(e.burkov):  Probably it's better to return the joined error from
+	// each upstream that returned no response, and get rid of multiple
+	// [errors.Is] calls.  This will change the behavior though.
 	if len(errs) == 0 {
-		// All responses had nil replies.
-		return nil, nil, fmt.Errorf("none of upstream servers responded")
+		return nil, nil, errors.Error("none of upstream servers responded")
 	}
 
-	return nil, nil, errors.List("all upstreams failed to respond", errs...)
+	return nil, nil, errors.Join(errs...)
 }
 
-// ExchangeAllResult - result of ExchangeAll()
+// ExchangeAllResult is the successful result of [ExchangeAll] for a single
+// upstream.
 type ExchangeAllResult struct {
-	Resp     *dns.Msg // response
-	Upstream Upstream // upstream server
+	// Resp is the response DNS request resolved into.
+	Resp *dns.Msg
+
+	// Upstream is the upstream that successfully resolved the request.
+	Upstream Upstream
 }
 
-// ExchangeAll receives a response from each of ups.
+// ExchangeAll returns the responses from all of u.  It returns an error only if
+// all upstreams failed to exchange the request.
 func ExchangeAll(ups []Upstream, req *dns.Msg) (res []ExchangeAllResult, err error) {
-	upsl := len(ups)
-	if upsl == 0 {
+	upsNum := len(ups)
+	switch upsNum {
+	case 0:
 		return nil, ErrNoUpstreams
+	case 1:
+		var reply *dns.Msg
+		reply, err = exchangeAndLog(ups[0], req)
+		if err != nil {
+			return nil, err
+		} else if reply == nil {
+			return nil, ErrNoReply
+		}
+
+		return []ExchangeAllResult{{Upstream: ups[0], Resp: reply}}, nil
+	default:
+		// Go on.
 	}
 
-	res = make([]ExchangeAllResult, 0, upsl)
-	errs := make([]error, 0, upsl)
-	resCh := make(chan *exchangeResult, upsl)
+	res = make([]ExchangeAllResult, 0, upsNum)
+	var errs []error
+
+	resCh := make(chan any, upsNum)
 
 	// Start exchanging concurrently.
 	for _, u := range ups {
@@ -86,138 +106,80 @@ func ExchangeAll(ups []Upstream, req *dns.Msg) (res []ExchangeAllResult, err err
 	}
 
 	// Wait for all exchanges to finish.
-	for i := 0; i < upsl; i++ {
-		rep := <-resCh
-		if rep.err != nil {
-			errs = append(errs, rep.err)
-
-			continue
+	for range ups {
+		var r *ExchangeAllResult
+		r, err = receiveAsyncResult(resCh)
+		if err != nil {
+			errs = append(errs, err)
+		} else {
+			res = append(res, *r)
 		}
-
-		if rep.reply == nil {
-			errs = append(errs, errors.Error("no reply"))
-
-			continue
-		}
-
-		res = append(res, ExchangeAllResult{
-			Resp:     rep.reply,
-			Upstream: rep.upstream,
-		})
 	}
-	if len(errs) == upsl {
+
+	if len(errs) == upsNum {
+		// TODO(e.burkov):  Use [errors.Join] in Go 1.20.
 		return res, errors.List("all upstreams failed to exchange", errs...)
 	}
 
-	return res, nil
+	return slices.Clip(res), nil
 }
 
-// exchangeAsync tries to resolve DNS request with one upstream and send result to resp channel
-func exchangeAsync(u Upstream, req *dns.Msg, respCh chan *exchangeResult) {
-	resp, err := u.Exchange(req.Copy())
-	respCh <- &exchangeResult{
-		reply:    resp,
-		upstream: u,
-		err:      err,
+// receiveAsyncResult receives a single result from resCh or an error from
+// errCh.  It returns either a non-nil result or an error.
+func receiveAsyncResult(resCh chan any) (res *ExchangeAllResult, err error) {
+	switch res := (<-resCh).(type) {
+	case error:
+		return nil, res
+	case *ExchangeAllResult:
+		if res.Resp == nil {
+			return nil, ErrNoReply
+		}
+
+		return res, nil
+	default:
+		return nil, fmt.Errorf("unexpected type %T of result", res)
 	}
 }
 
-func exchange(u Upstream, req *dns.Msg) (*dns.Msg, error) {
+// exchangeAsync tries to resolve DNS request with one upstream and sends the
+// result to respCh.
+func exchangeAsync(u Upstream, req *dns.Msg, resCh chan any) {
+	reply, err := exchangeAndLog(u, req)
+	if err != nil {
+		resCh <- err
+	} else {
+		resCh <- &ExchangeAllResult{Resp: reply, Upstream: u}
+	}
+}
+
+// exchangeAndLog wraps the [Upstream.Exchange] method with logging.
+func exchangeAndLog(u Upstream, req *dns.Msg) (resp *dns.Msg, err error) {
+	addr := u.Address()
+	req = req.Copy()
+
 	start := time.Now()
 	reply, err := u.Exchange(req)
-	elapsed := time.Since(start)
-	if err == nil {
-		log.Tracef(
-			"upstream %s successfully finished exchange of %s. Elapsed %s.",
-			u.Address(),
-			req.Question[0].String(),
-			elapsed,
-		)
-	} else {
-		log.Tracef(
-			"upstream %s failed to exchange %s in %s. Cause: %s",
-			u.Address(),
-			req.Question[0].String(),
-			elapsed,
-			err,
-		)
+	dur := time.Since(start)
+
+	if len(req.Question) > 0 {
+		if q := &req.Question[0]; err == nil {
+			log.Debug("dnsproxy: upstream %s exchanged %s successfully in %s", addr, q, dur)
+		} else {
+			log.Debug("dnsproxy: upstream %s failed to exchange %s in %s: %s", addr, q, dur, err)
+		}
 	}
+
 	return reply, err
 }
 
-// lookupResult is a structure that represents result of lookup
-type lookupResult struct {
-	address []net.IPAddr // List of IP addresses
-	err     error        // Error
-}
-
-// LookupParallel starts parallel lookup for host ip with many Resolvers
-// First answer without error will be returned
-// Return nil and error if count of errors equals count of resolvers
-func LookupParallel(ctx context.Context, resolvers []*Resolver, host string) ([]net.IPAddr, error) {
-	size := len(resolvers)
-
-	if size == 0 {
-		return nil, errors.Error("no resolvers specified")
-	}
-	if size == 1 {
-		address, err := lookup(ctx, resolvers[0], host)
-		return address, err
-	}
-
-	// Size of channel must accommodate results of lookups from all resolvers
-	// Otherwise sending in channel will be locked
-	ch := make(chan *lookupResult, size)
-
-	for _, res := range resolvers {
-		go lookupAsync(ctx, res, host, ch)
-	}
-
-	var errs []error
-	for n := 0; n < size; n++ {
-		result := <-ch
-
-		if result.err != nil {
-			errs = append(errs, result.err)
-
-			continue
-		}
-
-		return result.address, nil
-	}
-
-	return nil, errors.List("all resolvers failed", errs...)
-}
-
-// lookupAsync tries to lookup for host ip with one Resolver and sends lookupResult to res channel
-func lookupAsync(ctx context.Context, r *Resolver, host string, res chan *lookupResult) {
-	address, err := lookup(ctx, r, host)
-	res <- &lookupResult{
-		err:     err,
-		address: address,
-	}
-}
-
-func lookup(ctx context.Context, r *Resolver, host string) ([]net.IPAddr, error) {
-	start := time.Now()
-	address, err := r.LookupIPAddr(ctx, host)
-	elapsed := time.Since(start)
-	if err != nil {
-		log.Tracef(
-			"failed to lookup for %s in %s using %s: %s",
-			host,
-			elapsed,
-			r.resolverAddress,
-			err,
-		)
-	} else {
-		log.Tracef(
-			"successfully finished lookup for %s in %s using %s. Result : %s",
-			host,
-			elapsed,
-			r.resolverAddress,
-			address,
-		)
-	}
-	return address, err
+// LookupParallel tries to lookup for ip of host with all resolvers
+// concurrently.
+//
+// Deprecated:  Use [ParallelResolver] instead.
+func LookupParallel(
+	ctx context.Context,
+	resolvers []Resolver,
+	host string,
+) (addrs []netip.Addr, err error) {
+	return ParallelResolver(resolvers).LookupNetIP(ctx, "ip", host)
 }
