@@ -2,6 +2,7 @@ package fastip
 
 import (
 	"net"
+	"net/netip"
 	"runtime"
 	"sync"
 	"syscall"
@@ -28,8 +29,8 @@ func TestFastestAddr_PingAll_timeout(t *testing.T) {
 			return nil
 		}
 
-		ip := net.IP{127, 0, 0, 1}
-		res := f.pingAll("", []net.IP{ip, ip})
+		ip := netutil.IPv4Localhost()
+		res := f.pingAll("", []netip.Addr{ip, ip})
 		require.Nil(t, res)
 
 		waitCh <- unit{}
@@ -40,7 +41,8 @@ func TestFastestAddr_PingAll_timeout(t *testing.T) {
 
 		const lat uint = 42
 
-		ip1, ip2 := net.IP{127, 0, 0, 1}, net.IP{127, 0, 0, 2}
+		ip1 := netutil.IPv4Localhost()
+		ip2 := netip.MustParseAddr("127.0.0.2")
 		f.cacheAddSuccessful(ip1, lat)
 
 		waitCh := make(chan unit)
@@ -50,7 +52,7 @@ func TestFastestAddr_PingAll_timeout(t *testing.T) {
 			return nil
 		}
 
-		res := f.pingAll("", []net.IP{ip1, ip2})
+		res := f.pingAll("", []netip.Addr{ip1, ip2})
 		require.NotNil(t, res)
 
 		assert.True(t, res.success)
@@ -62,7 +64,7 @@ func TestFastestAddr_PingAll_timeout(t *testing.T) {
 
 // assertCaching checks the cache of f for containing a connection to ip with
 // the specified status.
-func assertCaching(t *testing.T, f *FastestAddr, ip net.IP, status int) {
+func assertCaching(t *testing.T, f *FastestAddr, ip netip.Addr, status int) {
 	t.Helper()
 
 	const tickDur = pingTCPTimeout / 16
@@ -75,23 +77,23 @@ func assertCaching(t *testing.T, f *FastestAddr, ip net.IP, status int) {
 }
 
 func TestFastestAddr_PingAll_cache(t *testing.T) {
-	ip := net.IP{127, 0, 0, 1}
+	ip := netutil.IPv4Localhost()
 
 	t.Run("cached_failed", func(t *testing.T) {
 		f := NewFastestAddr()
 		f.cacheAddFailure(ip)
 
-		res := f.pingAll("", []net.IP{ip, ip})
+		res := f.pingAll("", []netip.Addr{ip, ip})
 		require.Nil(t, res)
 	})
 
-	t.Run("cached_succesfull", func(t *testing.T) {
+	t.Run("cached_successful", func(t *testing.T) {
 		const lat uint = 1
 
 		f := NewFastestAddr()
 		f.cacheAddSuccessful(ip, lat)
 
-		res := f.pingAll("", []net.IP{ip, ip})
+		res := f.pingAll("", []netip.Addr{ip, ip})
 		require.NotNil(t, res)
 		assert.True(t, res.success)
 		assert.Equal(t, lat, res.latency)
@@ -102,16 +104,16 @@ func TestFastestAddr_PingAll_cache(t *testing.T) {
 		require.NoError(t, err)
 		testutil.CleanupAndRequireSuccess(t, listener.Close)
 
-		ip := net.IP{127, 0, 0, 1}
+		ip = netutil.IPv4Localhost()
 		f := NewFastestAddr()
 
 		f.pingPorts = []uint{uint(listener.Addr().(*net.TCPAddr).Port)}
-		ips := []net.IP{ip, ip}
+		ips := []netip.Addr{ip, ip}
 
 		wg := &sync.WaitGroup{}
 		wg.Add(len(ips) * len(f.pingPorts))
 
-		f.pinger.Control = func(_, address string, _ syscall.RawConn) error {
+		f.pinger.Control = func(_, address string, _ syscall.RawConn) (err error) {
 			hostport, err := netutil.ParseHostPort(address)
 			require.NoError(t, err)
 
@@ -134,10 +136,10 @@ func TestFastestAddr_PingAll_cache(t *testing.T) {
 }
 
 // listen is a helper function that creates a new listener on ip for t.
-func listen(t *testing.T, ip net.IP) (port uint) {
+func listen(t *testing.T, ip netip.Addr) (port uint) {
 	t.Helper()
 
-	l, err := net.Listen("tcp", netutil.IPPort{IP: ip, Port: 0}.String())
+	l, err := net.Listen("tcp", netip.AddrPortFrom(ip, 0).String())
 	require.NoError(t, err)
 	testutil.CleanupAndRequireSuccess(t, l.Close)
 
@@ -145,20 +147,20 @@ func listen(t *testing.T, ip net.IP) (port uint) {
 }
 
 func TestFastestAddr_PingAll(t *testing.T) {
-	ip := net.IP{127, 0, 0, 1}
+	ip := netutil.IPv4Localhost()
 
 	t.Run("single", func(t *testing.T) {
 		f := NewFastestAddr()
-		res := f.pingAll("", []net.IP{ip})
+		res := f.pingAll("", []netip.Addr{ip})
 		require.NotNil(t, res)
 
 		assert.True(t, res.success)
-		assert.True(t, ip.Equal(res.ipp.IP))
+		assert.Equal(t, ip, res.addrPort.Addr())
 		// There was no ping so the port is zero.
-		assert.Zero(t, res.ipp.Port)
+		assert.Zero(t, res.addrPort.Port())
 
 		// Nothing in the cache since there was no ping.
-		ce := f.cacheFind(res.ipp.IP)
+		ce := f.cacheFind(res.addrPort.Addr())
 		require.Nil(t, ce)
 	})
 
@@ -174,11 +176,9 @@ func TestFastestAddr_PingAll(t *testing.T) {
 			slowPort,
 		}
 		f.pinger.Control = func(_, address string, _ syscall.RawConn) error {
-			ipp, err := netutil.ParseIPPort(address)
-			require.NoError(t, err)
-
-			require.Contains(t, []uint{fastPort, slowPort}, uint(ipp.Port))
-			if ipp.Port == int(fastPort) {
+			addrPort := netip.MustParseAddrPort(address)
+			require.Contains(t, []uint{fastPort, slowPort}, uint(addrPort.Port()))
+			if addrPort.Port() == uint16(fastPort) {
 				return nil
 			}
 
@@ -187,16 +187,15 @@ func TestFastestAddr_PingAll(t *testing.T) {
 			return nil
 		}
 
-		ips := []net.IP{ip, ip}
-
+		ips := []netip.Addr{ip, ip}
 		res := f.pingAll("", ips)
 		ctrlCh <- unit{}
 
 		require.NotNil(t, res)
 
 		assert.True(t, res.success)
-		assert.True(t, ip.Equal(res.ipp.IP))
-		assert.EqualValues(t, fastPort, res.ipp.Port)
+		assert.Equal(t, ip, res.addrPort.Addr())
+		assert.EqualValues(t, fastPort, res.addrPort.Port())
 
 		assertCaching(t, f, ip, 0)
 	})
@@ -212,7 +211,7 @@ func TestFastestAddr_PingAll(t *testing.T) {
 		f := NewFastestAddr()
 		f.pingPorts = []uint{port}
 
-		res := f.pingAll("test", []net.IP{ip, ip})
+		res := f.pingAll("test", []netip.Addr{ip, ip})
 		require.Nil(t, res)
 
 		assertCaching(t, f, ip, 1)
