@@ -7,14 +7,15 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"strconv"
+	"net/netip"
+	"net/url"
 	"strings"
 
+	"github.com/AdguardTeam/golibs/httphdr"
 	"github.com/AdguardTeam/golibs/log"
-	"github.com/AdguardTeam/golibs/netutil"
-	"github.com/lucas-clemente/quic-go"
-	"github.com/lucas-clemente/quic-go/http3"
 	"github.com/miekg/dns"
+	"github.com/quic-go/quic-go"
+	"github.com/quic-go/quic-go/http3"
 	"golang.org/x/net/http2"
 )
 
@@ -70,9 +71,9 @@ func (p *Proxy) createHTTPSListeners() (err error) {
 	for _, addr := range p.HTTPSListenAddr {
 		log.Info("Creating an HTTPS server")
 
-		tcpAddr, err := p.listenHTTP(addr)
-		if err != nil {
-			return fmt.Errorf("failed to start HTTPS server on %s: %w", addr, err)
+		tcpAddr, lErr := p.listenHTTP(addr)
+		if lErr != nil {
+			return fmt.Errorf("failed to start HTTPS server on %s: %w", addr, lErr)
 		}
 
 		if p.HTTP3 {
@@ -97,169 +98,197 @@ func (p *Proxy) createHTTPSListeners() (err error) {
 //     "application/dns-message";
 //   - http.StatusMethodNotAllowed if request method is not GET or POST.
 func (p *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("Incoming HTTPS request on %s", r.URL)
+	log.Debug("dnsproxy: incoming https request on %s", r.URL)
+
+	raddr, prx, err := remoteAddr(r)
+	if err != nil {
+		log.Debug("dnsproxy: warning: getting real ip: %s", err)
+	}
+
+	if !p.checkBasicAuth(w, r, raddr) {
+		return
+	}
 
 	var buf []byte
-	var err error
 
 	switch r.Method {
 	case http.MethodGet:
 		dnsParam := r.URL.Query().Get("dns")
 		buf, err = base64.RawURLEncoding.DecodeString(dnsParam)
 		if len(buf) == 0 || err != nil {
-			log.Tracef("Cannot parse DNS request from %s", dnsParam)
+			log.Debug("dnsproxy: parsing dns request from get param %q: %v", dnsParam, err)
 			http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+
 			return
 		}
 	case http.MethodPost:
 		contentType := r.Header.Get("Content-Type")
 		if contentType != "application/dns-message" {
-			log.Tracef("Unsupported media type: %s", contentType)
+			log.Debug("dnsproxy: unsupported media type %q", contentType)
 			http.Error(w, http.StatusText(http.StatusUnsupportedMediaType), http.StatusUnsupportedMediaType)
+
 			return
 		}
 
 		buf, err = io.ReadAll(r.Body)
 		if err != nil {
-			log.Tracef("Cannot read the request body: %s", err)
+			log.Debug("dnsproxy: reading http request body: %s", err)
 			http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+
 			return
 		}
-		defer r.Body.Close()
+
+		defer log.OnCloserError(r.Body, log.DEBUG)
 	default:
-		log.Tracef("Wrong HTTP method: %s", r.Method)
+		log.Debug("dnsproxy: bad http method %q", r.Method)
 		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+
 		return
 	}
 
 	req := &dns.Msg{}
 	if err = req.Unpack(buf); err != nil {
-		log.Tracef("msg.Unpack: %s", err)
+		log.Debug("dnsproxy: unpacking http msg: %s", err)
 		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+
 		return
 	}
 
-	addr, prx, err := remoteAddr(r)
-	if err != nil {
-		log.Debug("warning: getting real ip: %s", err)
-	}
-
 	d := p.newDNSContext(ProtoHTTPS, req)
-	d.Addr = addr
+	d.Addr = raddr
 	d.HTTPRequest = r
 	d.HTTPResponseWriter = w
 
-	if prx != nil {
-		ip, _ := netutil.IPAndPortFromAddr(prx)
-		log.Debug("request came from proxy server %s", prx)
-		if !p.proxyVerifier.Contains(ip) {
-			log.Debug("proxy %s is not trusted, using original remote addr", ip)
+	if prx.IsValid() {
+		log.Debug("dnsproxy: request came from proxy server %s", prx)
+
+		// TODO(s.chzhen):  Consider using []netip.Prefix.
+		if !p.proxyVerifier.Contains(prx.Addr().AsSlice()) {
+			log.Debug("dnsproxy: proxy %s is not trusted, using original remote addr", prx)
 			d.Addr = prx
 		}
 	}
 
 	err = p.handleDNSRequest(d)
 	if err != nil {
-		log.Tracef("error handling DNS (%s) request: %s", d.Proto, err)
+		log.Debug("dnsproxy: handling dns (%s) request: %s", d.Proto, err)
 	}
 }
 
+// checkBasicAuth checks the basic authorization data, if necessary, and if the
+// data isn't valid, it writes an error.  shouldHandle is false if the request
+// has been denied.
+func (p *Proxy) checkBasicAuth(
+	w http.ResponseWriter,
+	r *http.Request,
+	raddr netip.AddrPort,
+) (shouldHandle bool) {
+	ui := p.Config.Userinfo
+	if ui == nil {
+		return true
+	}
+
+	user, pass, _ := r.BasicAuth()
+	if matchesUserinfo(ui, user, pass) {
+		return true
+	}
+
+	log.Error("dnsproxy: basic auth failed for user %q from raddr %s", user, raddr)
+
+	h := w.Header()
+	// TODO(a.garipov): Add to httphdr.
+	h.Set("Www-Authenticate", `Basic realm="DNS", charset="UTF-8"`)
+	http.Error(w, "Authorization required", http.StatusUnauthorized)
+
+	return false
+}
+
+// matchesUserinfo returns false if user and pass don't match userinfo.
+// userinfo must not be nil.
+func matchesUserinfo(userinfo *url.Userinfo, user, pass string) (ok bool) {
+	requiredPassword, _ := userinfo.Password()
+
+	return user == userinfo.Username() && pass == requiredPassword
+}
+
 // Writes a response to the DoH client.
-func (p *Proxy) respondHTTPS(d *DNSContext) error {
+func (p *Proxy) respondHTTPS(d *DNSContext) (err error) {
 	resp := d.Res
 	w := d.HTTPResponseWriter
 
 	if resp == nil {
-		// If no response has been written, indicate it via a 500 error
+		// Indicate the response's absence via a http.StatusInternalServerError.
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+
 		return nil
 	}
 
 	bytes, err := resp.Pack()
 	if err != nil {
-
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
 
 		return fmt.Errorf("packing message: %w", err)
 	}
 
-	w.Header().Set("Server", "AdGuard DNS")
-	w.Header().Set("Content-Type", "application/dns-message")
+	if srvName := p.Config.HTTPSServerName; srvName != "" {
+		w.Header().Set(httphdr.Server, srvName)
+	}
+
+	w.Header().Set(httphdr.ContentType, "application/dns-message")
 	_, err = w.Write(bytes)
 
 	return err
 }
 
 // realIPFromHdrs extracts the actual client's IP address from the first
-// suitable r's header.  It returns nil if r doesn't contain any information
-// about real client's IP address.  Current headers priority is:
+// suitable r's header.  It returns an error if r doesn't contain any
+// information about real client's IP address.  Current headers priority is:
 //
-//  1. CF-Connecting-IP
-//  2. True-Client-IP
-//  3. X-Real-IP
-//  4. X-Forwarded-For
-func realIPFromHdrs(r *http.Request) (realIP net.IP) {
+//  1. [httphdr.CFConnectingIP]
+//  2. [httphdr.TrueClientIP]
+//  3. [httphdr.XRealIP]
+//  4. [httphdr.XForwardedFor]
+func realIPFromHdrs(r *http.Request) (realIP netip.Addr, err error) {
 	for _, h := range []string{
-		// Headers set by CloudFlare proxy servers.
-		"CF-Connecting-IP",
-		"True-Client-IP",
-		// Other proxying headers.
-		"X-Real-IP",
+		httphdr.CFConnectingIP,
+		httphdr.TrueClientIP,
+		httphdr.XRealIP,
 	} {
-		realIP = net.ParseIP(strings.TrimSpace(r.Header.Get(h)))
-		if realIP != nil {
-			return realIP
+		realIP, err = netip.ParseAddr(strings.TrimSpace(r.Header.Get(h)))
+		if err == nil {
+			return realIP, nil
 		}
 	}
 
-	xff := r.Header.Get("X-Forwarded-For")
+	xff := r.Header.Get(httphdr.XForwardedFor)
 	firstComma := strings.IndexByte(xff, ',')
-	if firstComma == -1 {
-		return net.ParseIP(strings.TrimSpace(xff))
+	if firstComma > 0 {
+		xff = xff[:firstComma]
 	}
 
-	return net.ParseIP(strings.TrimSpace(xff[:firstComma]))
+	return netip.ParseAddr(strings.TrimSpace(xff))
 }
 
 // remoteAddr returns the real client's address and the IP address of the latest
 // proxy server if any.
-func remoteAddr(r *http.Request) (addr, prx net.Addr, err error) {
-	var hostStr, portStr string
-	if hostStr, portStr, err = net.SplitHostPort(r.RemoteAddr); err != nil {
-		return nil, nil, err
+func remoteAddr(r *http.Request) (addr, prx netip.AddrPort, err error) {
+	host, err := netip.ParseAddrPort(r.RemoteAddr)
+	if err != nil {
+		return netip.AddrPort{}, netip.AddrPort{}, err
 	}
 
-	var port int
-	if port, err = strconv.Atoi(portStr); err != nil {
-		return nil, nil, err
+	realIP, err := realIPFromHdrs(r)
+	if err != nil {
+		log.Debug("dnsproxy: getting ip address from http request: %s", err)
+
+		return host, netip.AddrPort{}, nil
 	}
 
-	host := net.ParseIP(hostStr)
-	if host == nil {
-		return nil, nil, fmt.Errorf("invalid ip: %s", hostStr)
-	}
+	log.Debug("dnsproxy: using ip address from http request: %s", realIP)
 
-	h3 := r.Context().Value(http3.ServerContextKey) != nil
+	// TODO(a.garipov): Add port if we can get it from headers like X-Real-Port,
+	// X-Forwarded-Port, etc.
+	addr = netip.AddrPortFrom(realIP, 0)
 
-	if realIP := realIPFromHdrs(r); realIP != nil {
-		log.Tracef("Using IP address from HTTP request: %s", realIP)
-
-		// TODO(a.garipov): Add port if we can get it from headers like
-		// X-Real-Port, X-Forwarded-Port, etc.
-		if h3 {
-			addr = &net.UDPAddr{IP: realIP, Port: 0}
-			prx = &net.UDPAddr{IP: host, Port: port}
-		} else {
-			addr = &net.TCPAddr{IP: realIP, Port: 0}
-			prx = &net.TCPAddr{IP: host, Port: port}
-		}
-
-		return addr, prx, nil
-	}
-
-	if h3 {
-		return &net.UDPAddr{IP: host, Port: port}, nil, nil
-	}
-
-	return &net.TCPAddr{IP: host, Port: port}, nil, nil
+	return addr, host, nil
 }
