@@ -9,15 +9,15 @@ import (
 	"io"
 	"net"
 	"net/http"
+	"net/netip"
 	"net/url"
 	"strings"
 	"testing"
 
-	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/AdguardTeam/golibs/testutil"
-	"github.com/lucas-clemente/quic-go"
-	"github.com/lucas-clemente/quic-go/http3"
 	"github.com/miekg/dns"
+	"github.com/quic-go/quic-go"
+	"github.com/quic-go/quic-go/http3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -61,16 +61,19 @@ func TestHttpsProxy(t *testing.T) {
 }
 
 func TestProxy_trustedProxies(t *testing.T) {
-	clientIP, proxyIP := net.IP{1, 2, 3, 4}, net.IP{127, 0, 0, 1}
+	var (
+		clientAddr = netip.MustParseAddr("1.2.3.4")
+		proxyAddr  = netip.MustParseAddr("127.0.0.1")
+	)
 
-	doRequest := func(t *testing.T, proxyAddr string, expectedClientIP net.IP) {
+	doRequest := func(t *testing.T, addr string, expectedClientIP netip.Addr) {
 		// Prepare the proxy server.
 		tlsConf, caPem := createServerTLSConfig(t)
 		dnsProxy := createTestProxy(t, tlsConf)
 
-		var gotAddr net.Addr
+		var gotAddr netip.Addr
 		dnsProxy.RequestHandler = func(_ *Proxy, d *DNSContext) (err error) {
-			gotAddr = d.Addr
+			gotAddr = d.Addr.Addr()
 
 			return dnsProxy.Resolve(d)
 		}
@@ -79,7 +82,7 @@ func TestProxy_trustedProxies(t *testing.T) {
 
 		msg := createTestMessage()
 
-		dnsProxy.TrustedProxies = []string{proxyAddr}
+		dnsProxy.TrustedProxies = []string{addr}
 
 		// Start listening.
 		serr := dnsProxy.Start()
@@ -87,51 +90,59 @@ func TestProxy_trustedProxies(t *testing.T) {
 		testutil.CleanupAndRequireSuccess(t, dnsProxy.Stop)
 
 		hdrs := map[string]string{
-			"X-Forwarded-For": strings.Join([]string{clientIP.String(), proxyIP.String()}, ","),
+			"X-Forwarded-For": strings.Join([]string{clientAddr.String(), proxyAddr.String()}, ","),
 		}
 
 		resp := sendTestDoHMessage(t, client, msg, hdrs)
 		requireResponse(t, msg, resp)
 
-		ip, _ := netutil.IPAndPortFromAddr(gotAddr)
-		require.True(t, ip.Equal(expectedClientIP))
+		require.Equal(t, expectedClientIP, gotAddr)
 	}
 
 	t.Run("success", func(t *testing.T) {
-		doRequest(t, proxyIP.String(), clientIP)
+		doRequest(t, proxyAddr.String(), clientAddr)
 	})
 
 	t.Run("not_in_trusted", func(t *testing.T) {
-		doRequest(t, "127.0.0.2", proxyIP)
+		doRequest(t, "127.0.0.2", proxyAddr)
 	})
 }
 
 func TestAddrsFromRequest(t *testing.T) {
-	theIP, anotherIP := net.IP{1, 2, 3, 4}, net.IP{1, 2, 3, 5}
-	theIPStr, anotherIPStr := theIP.String(), anotherIP.String()
+	var (
+		theIP     = netip.AddrFrom4([4]byte{1, 2, 3, 4})
+		anotherIP = netip.AddrFrom4([4]byte{1, 2, 3, 5})
+
+		theIPStr     = theIP.String()
+		anotherIPStr = anotherIP.String()
+	)
 
 	testCases := []struct {
-		name   string
-		hdrs   map[string]string
-		wantIP net.IP
+		name    string
+		hdrs    map[string]string
+		wantIP  netip.Addr
+		wantErr string
 	}{{
 		name: "cf-connecting-ip",
 		hdrs: map[string]string{
 			"CF-Connecting-IP": theIPStr,
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "true-client-ip",
 		hdrs: map[string]string{
 			"True-Client-IP": theIPStr,
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "x-real-ip",
 		hdrs: map[string]string{
 			"X-Real-IP": theIPStr,
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "no_any",
 		hdrs: map[string]string{
@@ -139,7 +150,8 @@ func TestAddrsFromRequest(t *testing.T) {
 			"True-Client-IP":   "invalid",
 			"X-Real-IP":        "invalid",
 		},
-		wantIP: nil,
+		wantIP:  netip.Addr{},
+		wantErr: `ParseAddr(""): unable to parse IP`,
 	}, {
 		name: "priority",
 		hdrs: map[string]string{
@@ -148,47 +160,54 @@ func TestAddrsFromRequest(t *testing.T) {
 			"X-Real-IP":        anotherIPStr,
 			"CF-Connecting-IP": theIPStr,
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "x-forwarded-for_simple",
 		hdrs: map[string]string{
 			"X-Forwarded-For": strings.Join([]string{anotherIPStr, theIPStr}, ","),
 		},
-		wantIP: anotherIP,
+		wantIP:  anotherIP,
+		wantErr: "",
 	}, {
 		name: "x-forwarded-for_single",
 		hdrs: map[string]string{
 			"X-Forwarded-For": theIPStr,
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "x-forwarded-for_invalid_proxy",
 		hdrs: map[string]string{
 			"X-Forwarded-For": strings.Join([]string{theIPStr, "invalid"}, ","),
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "x-forwarded-for_empty",
 		hdrs: map[string]string{
 			"X-Forwarded-For": "",
 		},
-		wantIP: nil,
+		wantIP:  netip.Addr{},
+		wantErr: `ParseAddr(""): unable to parse IP`,
 	}, {
 		name: "x-forwarded-for_redundant_spaces",
 		hdrs: map[string]string{
 			"X-Forwarded-For": "  " + theIPStr + "   ,\t" + anotherIPStr,
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}, {
 		name: "cf-connecting-ip_redundant_spaces",
 		hdrs: map[string]string{
 			"CF-Connecting-IP": "  " + theIPStr + "\t",
 		},
-		wantIP: theIP,
+		wantIP:  theIP,
+		wantErr: "",
 	}}
 
 	for _, tc := range testCases {
-		r, err := http.NewRequest("GET", "localhost", nil)
+		r, err := http.NewRequest(http.MethodGet, "localhost", nil)
 		require.NoError(t, err)
 
 		for h, v := range tc.hdrs {
@@ -196,31 +215,44 @@ func TestAddrsFromRequest(t *testing.T) {
 		}
 
 		t.Run(tc.name, func(t *testing.T) {
-			ip := realIPFromHdrs(r)
-			assert.True(t, tc.wantIP.Equal(ip))
+			var ip netip.Addr
+			ip, err = realIPFromHdrs(r)
+			testutil.AssertErrorMsg(t, tc.wantErr, err)
+
+			assert.Equal(t, tc.wantIP, ip)
 		})
 	}
 }
 
 func TestRemoteAddr(t *testing.T) {
-	theIP, anotherIP, thirdIP := net.IP{1, 2, 3, 4}, net.IP{1, 2, 3, 5}, net.IP{1, 2, 3, 6}
-	theIPStr, anotherIPStr, thirdIPStr := theIP.String(), anotherIP.String(), thirdIP.String()
-	rAddr := &net.TCPAddr{IP: theIP, Port: 1}
+	const thePort = 4321
+
+	var (
+		theIP     = netip.AddrFrom4([4]byte{1, 2, 3, 4})
+		anotherIP = netip.AddrFrom4([4]byte{1, 2, 3, 5})
+		thirdIP   = netip.AddrFrom4([4]byte{1, 2, 3, 6})
+
+		theIPStr     = theIP.String()
+		anotherIPStr = anotherIP.String()
+		thirdIPStr   = thirdIP.String()
+	)
+
+	rAddr := netip.AddrPortFrom(theIP, thePort)
 
 	testCases := []struct {
 		name       string
 		remoteAddr string
 		hdrs       map[string]string
 		wantErr    string
-		wantIP     net.IP
-		wantProxy  net.IP
+		wantIP     netip.AddrPort
+		wantProxy  netip.AddrPort
 	}{{
 		name:       "no_proxy",
 		remoteAddr: rAddr.String(),
 		hdrs:       nil,
 		wantErr:    "",
-		wantIP:     theIP,
-		wantProxy:  nil,
+		wantIP:     netip.AddrPortFrom(theIP, thePort),
+		wantProxy:  netip.AddrPort{},
 	}, {
 		name:       "proxied_with_cloudflare",
 		remoteAddr: rAddr.String(),
@@ -228,8 +260,8 @@ func TestRemoteAddr(t *testing.T) {
 			"CF-Connecting-IP": anotherIPStr,
 		},
 		wantErr:   "",
-		wantIP:    anotherIP,
-		wantProxy: theIP,
+		wantIP:    netip.AddrPortFrom(anotherIP, 0),
+		wantProxy: netip.AddrPortFrom(theIP, thePort),
 	}, {
 		name:       "proxied_once",
 		remoteAddr: rAddr.String(),
@@ -237,8 +269,8 @@ func TestRemoteAddr(t *testing.T) {
 			"X-Forwarded-For": anotherIPStr,
 		},
 		wantErr:   "",
-		wantIP:    anotherIP,
-		wantProxy: theIP,
+		wantIP:    netip.AddrPortFrom(anotherIP, 0),
+		wantProxy: netip.AddrPortFrom(theIP, thePort),
 	}, {
 		name:       "proxied_multiple",
 		remoteAddr: rAddr.String(),
@@ -246,42 +278,42 @@ func TestRemoteAddr(t *testing.T) {
 			"X-Forwarded-For": strings.Join([]string{anotherIPStr, thirdIPStr}, ","),
 		},
 		wantErr:   "",
-		wantIP:    anotherIP,
-		wantProxy: theIP,
+		wantIP:    netip.AddrPortFrom(anotherIP, 0),
+		wantProxy: netip.AddrPortFrom(theIP, thePort),
 	}, {
 		name:       "no_port",
 		remoteAddr: theIPStr,
 		hdrs:       nil,
-		wantErr:    "address " + theIPStr + ": missing port in address",
-		wantIP:     nil,
-		wantProxy:  nil,
+		wantErr:    "not an ip:port",
+		wantIP:     netip.AddrPort{},
+		wantProxy:  netip.AddrPort{},
 	}, {
 		name:       "bad_port",
 		remoteAddr: theIPStr + ":notport",
 		hdrs:       nil,
-		wantErr:    "strconv.Atoi: parsing \"notport\": invalid syntax",
-		wantIP:     nil,
-		wantProxy:  nil,
+		wantErr:    `invalid port "notport" parsing "1.2.3.4:notport"`,
+		wantIP:     netip.AddrPort{},
+		wantProxy:  netip.AddrPort{},
 	}, {
 		name:       "bad_host",
 		remoteAddr: "host:1",
 		hdrs:       nil,
-		wantErr:    "invalid ip: host",
-		wantIP:     nil,
-		wantProxy:  nil,
+		wantErr:    `ParseAddr("host"): unable to parse IP`,
+		wantIP:     netip.AddrPort{},
+		wantProxy:  netip.AddrPort{},
 	}, {
 		name:       "bad_proxied_host",
 		remoteAddr: "host:1",
 		hdrs: map[string]string{
 			"CF-Connecting-IP": theIPStr,
 		},
-		wantErr:   "invalid ip: host",
-		wantIP:    nil,
-		wantProxy: nil,
+		wantErr:   `ParseAddr("host"): unable to parse IP`,
+		wantIP:    netip.AddrPort{},
+		wantProxy: netip.AddrPort{},
 	}}
 
 	for _, tc := range testCases {
-		r, err := http.NewRequest("GET", "localhost", nil)
+		r, err := http.NewRequest(http.MethodGet, "localhost", nil)
 		require.NoError(t, err)
 
 		r.RemoteAddr = tc.remoteAddr
@@ -290,20 +322,17 @@ func TestRemoteAddr(t *testing.T) {
 		}
 
 		t.Run(tc.name, func(t *testing.T) {
-			addr, prx, err := remoteAddr(r)
+			var addr, prx netip.AddrPort
+			addr, prx, err = remoteAddr(r)
 			if tc.wantErr != "" {
-				assert.Equal(t, tc.wantErr, err.Error())
+				testutil.AssertErrorMsg(t, tc.wantErr, err)
 
 				return
 			}
 
 			require.NoError(t, err)
-
-			ip, _ := netutil.IPAndPortFromAddr(addr)
-			assert.True(t, ip.Equal(tc.wantIP))
-
-			prxIP, _ := netutil.IPAndPortFromAddr(prx)
-			assert.True(t, tc.wantProxy.Equal(prxIP))
+			assert.Equal(t, tc.wantIP, addr)
+			assert.Equal(t, tc.wantProxy, prx)
 		})
 	}
 }
@@ -387,7 +416,7 @@ func createTestHTTPClient(dnsProxy *Proxy, caPem []byte, http3Enabled bool) (cli
 				cfg *quic.Config,
 			) (quic.EarlyConnection, error) {
 				addr := dnsProxy.Addr(ProtoHTTPS).String()
-				return quic.DialAddrEarlyContext(ctx, addr, tlsCfg, cfg)
+				return quic.DialAddrEarly(ctx, addr, tlsCfg, cfg)
 			},
 			TLSClientConfig:    tlsClientConfig,
 			QuicConfig:         &quic.Config{},
