@@ -1,30 +1,38 @@
 package proxy
 
 import (
+	"context"
 	"crypto/tls"
+	"encoding/binary"
 	"fmt"
 	"io"
 	"net"
 	"time"
 
-	"github.com/AdguardTeam/dnsproxy/proxyutil"
+	proxynetutil "github.com/AdguardTeam/dnsproxy/internal/netutil"
 	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/log"
+	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/miekg/dns"
 )
 
-func (p *Proxy) createTCPListeners() (err error) {
+func (p *Proxy) createTCPListeners(ctx context.Context) (err error) {
 	for _, a := range p.TCPListenAddr {
-		log.Printf("Creating a TCP server socket")
+		log.Info("dnsproxy: creating tcp server socket %s", a)
 
-		var tcpListen *net.TCPListener
-		tcpListen, err = net.ListenTCP("tcp", a)
-		if err != nil {
-			return fmt.Errorf("starting listening on tcp socket: %w", err)
+		lsnr, lErr := proxynetutil.ListenConfig().Listen(ctx, "tcp", a.String())
+		if lErr != nil {
+			return fmt.Errorf("listening to tcp socket: %w", lErr)
 		}
 
-		p.tcpListen = append(p.tcpListen, tcpListen)
-		log.Printf("Listening to tcp://%s", tcpListen.Addr())
+		tcpListener, ok := lsnr.(*net.TCPListener)
+		if !ok {
+			return fmt.Errorf("wrong listener type on tcp addr %s: %T", a, lsnr)
+		}
+
+		p.tcpListen = append(p.tcpListen, tcpListener)
+
+		log.Info("dnsproxy: listening to tcp://%s", tcpListener.Addr())
 	}
 
 	return nil
@@ -32,17 +40,18 @@ func (p *Proxy) createTCPListeners() (err error) {
 
 func (p *Proxy) createTLSListeners() (err error) {
 	for _, a := range p.TLSListenAddr {
-		log.Printf("Creating a TLS server socket")
+		log.Info("dnsproxy: creating tls server socket %s", a)
 
 		var tcpListen *net.TCPListener
 		tcpListen, err = net.ListenTCP("tcp", a)
 		if err != nil {
-			return fmt.Errorf("starting tls listener: %w", err)
+			return fmt.Errorf("listening on tls addr %s: %w", a, err)
 		}
 
 		l := tls.NewListener(tcpListen, p.TLSConfig)
 		p.tlsListen = append(p.tlsListen, l)
-		log.Printf("Listening to tls://%s", l.Addr())
+
+		log.Info("dnsproxy: listening to tls://%s", l.Addr())
 	}
 
 	return nil
@@ -53,25 +62,25 @@ func (p *Proxy) createTLSListeners() (err error) {
 //
 // See also the comment on Proxy.requestGoroutinesSema.
 func (p *Proxy) tcpPacketLoop(l net.Listener, proto Proto, requestGoroutinesSema semaphore) {
-	log.Printf("Entering the %s listener loop on %s", proto, l.Addr())
+	log.Info("dnsproxy: entering %s listener loop on %s", proto, l.Addr())
+
 	for {
 		clientConn, err := l.Accept()
-
 		if err != nil {
 			if errors.Is(err, net.ErrClosed) {
-				log.Debug("tcpPacketLoop: connection closed: %s", err)
+				log.Debug("dnsproxy: tcp connection %s closed", l.Addr())
 			} else {
-				log.Info("got error when reading from TCP listen: %s", err)
+				log.Error("dnsproxy: reading from tcp: %s", err)
 			}
 
 			break
-		} else {
-			requestGoroutinesSema.acquire()
-			go func() {
-				p.handleTCPConnection(clientConn, proto)
-				requestGoroutinesSema.release()
-			}()
 		}
+
+		requestGoroutinesSema.acquire()
+		go func() {
+			p.handleTCPConnection(clientConn, proto)
+			requestGoroutinesSema.release()
+		}()
 	}
 }
 
@@ -80,11 +89,12 @@ func (p *Proxy) tcpPacketLoop(l net.Listener, proto Proto, requestGoroutinesSema
 func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
 	defer log.OnPanic("proxy.handleTCPConnection")
 
-	log.Tracef("handling tcp: started handling %s request from %s", proto, conn.RemoteAddr())
+	log.Debug("dnsproxy: handling new %s request from %s", proto, conn.RemoteAddr())
+
 	defer func() {
 		err := conn.Close()
 		if err != nil {
-			logWithNonCrit(err, "handling tcp: closing conn")
+			logWithNonCrit(err, "dnsproxy: handling tcp: closing conn")
 		}
 	}()
 
@@ -101,7 +111,7 @@ func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
 			logWithNonCrit(err, "handling tcp: setting deadline")
 		}
 
-		packet, err := proxyutil.ReadPrefixed(conn)
+		packet, err := readPrefixed(conn)
 		if err != nil {
 			logWithNonCrit(err, "handling tcp: reading msg")
 
@@ -111,13 +121,13 @@ func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
 		req := &dns.Msg{}
 		err = req.Unpack(packet)
 		if err != nil {
-			log.Error("handling tcp: unpacking msg: %s", err)
+			log.Error("dnsproxy: handling tcp: unpacking msg: %s", err)
 
 			return
 		}
 
 		d := p.newDNSContext(proto, req)
-		d.Addr = conn.RemoteAddr()
+		d.Addr = netutil.NetAddrToAddrPort(conn.RemoteAddr())
 		d.Conn = conn
 
 		err = p.handleDNSRequest(d)
@@ -125,6 +135,32 @@ func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
 			logWithNonCrit(err, fmt.Sprintf("handling tcp: handling %s request", d.Proto))
 		}
 	}
+}
+
+// errTooLarge means that a DNS message is larger than 64KiB.
+const errTooLarge errors.Error = "dns message is too large"
+
+// readPrefixed reads a DNS message with a 2-byte prefix containing message
+// length from conn.
+func readPrefixed(conn net.Conn) (b []byte, err error) {
+	l := make([]byte, 2)
+	_, err = conn.Read(l)
+	if err != nil {
+		return nil, fmt.Errorf("reading len: %w", err)
+	}
+
+	packetLen := binary.BigEndian.Uint16(l)
+	if packetLen > dns.MaxMsgSize {
+		return nil, errTooLarge
+	}
+
+	b = make([]byte, packetLen)
+	_, err = io.ReadFull(conn, b)
+	if err != nil {
+		return nil, fmt.Errorf("reading msg: %w", err)
+	}
+
+	return b, nil
 }
 
 // logWithNonCrit logs the error on the appropriate level depending on whether
@@ -154,10 +190,20 @@ func (p *Proxy) respondTCP(d *DNSContext) error {
 		return fmt.Errorf("packing message: %w", err)
 	}
 
-	err = proxyutil.WritePrefixed(bytes, conn)
+	err = writePrefixed(bytes, conn)
 	if err != nil && !errors.Is(err, net.ErrClosed) {
 		return fmt.Errorf("writing message: %w", err)
 	}
 
 	return nil
+}
+
+// writePrefixed writes a DNS message to a TCP connection it first writes
+// a 2-byte prefix followed by the message itself.
+func writePrefixed(b []byte, conn net.Conn) (err error) {
+	l := make([]byte, 2)
+	binary.BigEndian.PutUint16(l, uint16(len(b)))
+	_, err = (&net.Buffers{l, b}).WriteTo(conn)
+
+	return err
 }
