@@ -9,29 +9,160 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestGetUpstreamsForDomain(t *testing.T) {
+func TestUpstreamConfig_GetUpstreamsForDomain(t *testing.T) {
 	upstreams := []string{
 		"[/google.com/local/]4.3.2.1",
 		"[/www.google.com//]1.2.3.4",
 		"[/maps.google.com/]#",
 		"[/www.google.com/]tls://1.1.1.1",
+		"[/_acme-challenge.example.org/]#",
+		"[/example.com/]1.1.1.1 2.2.2.2 3.3.3.3",
 	}
 
 	config, err := ParseUpstreamsConfig(
 		upstreams,
 		&upstream.Options{
 			InsecureSkipVerify: false,
-			Bootstrap:          []string{},
+			Bootstrap:          nil,
 			Timeout:            1 * time.Second,
 		},
 	)
 	require.NoError(t, err)
 
-	assertUpstreamsForDomain(t, config, "www.google.com.", []string{"1.2.3.4:53", "tls://1.1.1.1:853"})
-	assertUpstreamsForDomain(t, config, "www2.google.com.", []string{"4.3.2.1:53"})
-	assertUpstreamsForDomain(t, config, "internal.local.", []string{"4.3.2.1:53"})
-	assertUpstreamsForDomain(t, config, "google.", []string{"1.2.3.4:53"})
-	assertUpstreamsForDomain(t, config, "maps.google.com.", []string{})
+	testCases := []struct {
+		name string
+		in   string
+		want []string
+	}{{
+		name: "direct_match",
+		in:   "www.google.com.",
+		want: []string{"1.2.3.4:53", "tls://1.1.1.1:853"},
+	}, {
+		name: "subdomain_match",
+		in:   "www2.google.com.",
+		want: []string{"4.3.2.1:53"},
+	}, {
+		name: "second_subdomain_match",
+		in:   "internal.local.",
+		want: []string{"4.3.2.1:53"},
+	}, {
+		name: "unqualified_match",
+		in:   "google.",
+		want: []string{"1.2.3.4:53"},
+	}, {
+		name: "default",
+		in:   "_acme-challenge.example.org.",
+		want: []string{},
+	}, {
+		name: "another_default",
+		in:   "maps.google.com.",
+		want: []string{},
+	}, {
+		name: "multiple_reserved",
+		in:   "example.com.",
+		want: []string{"1.1.1.1:53", "2.2.2.2:53", "3.3.3.3:53"},
+	}}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			ups := config.getUpstreamsForDomain(tc.in)
+			assertUpstreamsForDomain(t, ups, tc.want)
+		})
+	}
+}
+
+func TestUpstreamConfig_GetUpstreamsForDS(t *testing.T) {
+	upstreams := []string{
+		"[/google.com/local/]4.3.2.1",
+		"[/www.google.com//]1.2.3.4",
+		"[/maps.google.com/]#",
+		"[/www.google.com/]tls://1.1.1.1",
+		"[/_acme-challenge.example.org/]#",
+	}
+
+	config, err := ParseUpstreamsConfig(
+		upstreams,
+		&upstream.Options{
+			InsecureSkipVerify: false,
+			Timeout:            1 * time.Second,
+		},
+	)
+	require.NoError(t, err)
+
+	testCases := []struct {
+		name string
+		in   string
+		want []string
+	}{{
+		name: "direct_match",
+		in:   "www.google.com.",
+		want: []string{"4.3.2.1:53"},
+	}, {
+		name: "subdomain_match",
+		in:   "abc.www.google.com.",
+		want: []string{"1.2.3.4:53", "tls://1.1.1.1:853"},
+	}, {
+		name: "unqualified_match",
+		in:   "internal.local.",
+		want: []string{"1.2.3.4:53"},
+	}, {
+		name: "more_unqualified_match",
+		in:   "google.",
+		want: []string{"1.2.3.4:53"},
+	}, {
+		name: "default",
+		in:   "_acme-challenge.example.org.",
+		want: []string{},
+	}}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			ups := config.getUpstreamsForDS(tc.in)
+			assertUpstreamsForDomain(t, ups, tc.want)
+		})
+	}
+}
+
+func TestUpstreamConfig_Validate(t *testing.T) {
+	testCases := []struct {
+		name            string
+		wantValidateErr error
+		in              []string
+	}{{
+		name:            "empty",
+		wantValidateErr: upstream.ErrNoUpstreams,
+		in:              []string{},
+	}, {
+		name:            "nil",
+		wantValidateErr: upstream.ErrNoUpstreams,
+		in:              nil,
+	}, {
+		name:            "valid",
+		wantValidateErr: nil,
+		in: []string{
+			"udp://upstream.example:53",
+		},
+	}, {
+		name:            "no_default",
+		wantValidateErr: errNoDefaultUpstreams,
+		in: []string{
+			"[/domain.example/]udp://upstream.example:53",
+			"[/another.domain.example/]#",
+		},
+	}}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			c, err := ParseUpstreamsConfig(tc.in, nil)
+			require.NoError(t, err)
+
+			assert.ErrorIs(t, c.validate(), tc.wantValidateErr)
+		})
+	}
+
+	t.Run("actual_nil", func(t *testing.T) {
+		assert.ErrorIs(t, (*UpstreamConfig)(nil).validate(), errNoDefaultUpstreams)
+	})
 }
 
 func TestGetUpstreamsForDomainWithoutDuplicates(t *testing.T) {
@@ -40,7 +171,7 @@ func TestGetUpstreamsForDomainWithoutDuplicates(t *testing.T) {
 		upstreams,
 		&upstream.Options{
 			InsecureSkipVerify: false,
-			Bootstrap:          []string{},
+			Bootstrap:          nil,
 			Timeout:            1 * time.Second,
 		},
 	)
@@ -113,7 +244,8 @@ func TestGetUpstreamsForDomain_wildcards(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			assertUpstreamsForDomain(t, uconf, tc.in, tc.want)
+			ups := uconf.getUpstreamsForDomain(tc.in)
+			assertUpstreamsForDomain(t, ups, tc.want)
 		})
 	}
 }
@@ -153,7 +285,8 @@ func TestGetUpstreamsForDomain_sub_wildcards(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			assertUpstreamsForDomain(t, uconf, tc.in, tc.want)
+			ups := uconf.getUpstreamsForDomain(tc.in)
+			assertUpstreamsForDomain(t, ups, tc.want)
 		})
 	}
 }
@@ -194,10 +327,14 @@ func TestGetUpstreamsForDomain_default_wildcards(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			assertUpstreamsForDomain(t, uconf, tc.in, tc.want)
+			ups := uconf.getUpstreamsForDomain(tc.in)
+			assertUpstreamsForDomain(t, ups, tc.want)
 		})
 	}
 }
+
+// upsSink is the typed sink variable for the result of benchmarked function.
+var upsSink []upstream.Upstream
 
 func BenchmarkGetUpstreamsForDomain(b *testing.B) {
 	upstreams := []string{
@@ -211,29 +348,31 @@ func BenchmarkGetUpstreamsForDomain(b *testing.B) {
 		upstreams,
 		&upstream.Options{
 			InsecureSkipVerify: false,
-			Bootstrap:          []string{},
+			Bootstrap:          nil,
 			Timeout:            1 * time.Second,
 		},
 	)
 
-	for i := 0; i < b.N; i++ {
-		assertUpstreamsForDomain(b, config, "www.google.com.", []string{"1.2.3.4:53", "tls://1.1.1.1:853"})
-		assertUpstreamsForDomain(b, config, "www2.google.com.", []string{"4.3.2.1:53"})
-		assertUpstreamsForDomain(b, config, "internal.local.", []string{"4.3.2.1:53"})
-		assertUpstreamsForDomain(b, config, "google.", []string{"1.2.3.4:53"})
-		assertUpstreamsForDomain(b, config, "maps.google.com.", []string{})
+	domains := []string{
+		"www.google.com.",
+		"www2.google.com.",
+		"internal.local.",
+		"google.",
+		"maps.google.com.",
+	}
+
+	for i, l := 0, len(domains); i < b.N; i++ {
+		upsSink = config.getUpstreamsForDomain(domains[i%l])
 	}
 }
 
 // assertUpstreamsForDomain checks the addresses of the specified domain
 // upstreams and their number.
-func assertUpstreamsForDomain(t testing.TB, config *UpstreamConfig, domain string, address []string) {
-	t.Helper()
+func assertUpstreamsForDomain(tb testing.TB, ups []upstream.Upstream, want []string) {
+	tb.Helper()
 
-	u := config.getUpstreamsForDomain(domain)
-	require.Len(t, u, len(address))
-
-	for i, up := range u {
-		assert.Equalf(t, address[i], up.Address(), "bad upstream at index %d", i)
+	require.Len(tb, ups, len(want))
+	for i, up := range ups {
+		assert.Equalf(tb, want[i], up.Address(), "bad upstream at index %d", i)
 	}
 }
